fix: fix all -Werror format warnings and compilation errors across shell commands
- Fix format specifier mismatches: %u with unsigned long → %llu with unsigned long long
  or %u with unsigned int for small values (MAC bytes, IP octets, sizes)
- Fix %c with unsigned long → (int) for character output
- Fix %x/%d mismatches with unsigned long casts throughout cmd_*.c files
- Remove redundant/incorrect (unsigned long) casts on uint32_t values
- Remove unused variables (header_cols, p_len, brand_id, l1_assoc, l1_line)
- Mark unused functions with __attribute__((unused)) (ncdu_sort_children, draw_bar)
- Fix misleading indentation in else clauses (cmd_zcmp, cmd_zegrep, cmd_zfgrep, cmd_zgrep)
- Add missing braces around multi-statement else blocks
- Add vfs.h include for cmd_ncdu.c (vfs_readdir_names)
- Fix implicit-function-declaration: use fs_list_names instead of vfs_readdir
- Fix incompatible pointer type in args_to_argv (char *** vs char **)
- Implement actual write in cmd_shred.c (was a no-op before)
- Fix format string in test.c: unsigned long long vs unsigned long

All builds now pass with -Werror with zero warnings across 500+ C files.

+++ src/shell/cmds/cmd_fbinfo.c
@@ -12,13 +12,13 @@ void cmd_fbinfo(void) {
 
     if (!info.is_framebuffer) {
         kprintf("Display backend: VGA text mode\n");
-        kprintf("Console grid: %ux%u\n", (unsigned long)VGA_WIDTH, (unsigned long)VGA_HEIGHT);
+        kprintf("Console grid: %ux%u\n", VGA_WIDTH, VGA_HEIGHT);
         return;
     }
 
     kprintf("Display backend: framebuffer\n");
     kprintf("Framebuffer: %ux%u, %u bpp, pitch=%u bytes\n",
-            (uint64_t)info.width, (uint64_t)info.height,
-            (uint64_t)info.bpp, (uint64_t)info.pitch);
-    kprintf("Console grid: %ux%u cells\n", (unsigned long)VGA_WIDTH, (unsigned long)VGA_HEIGHT);
+            (unsigned int)info.width, (unsigned int)info.height,
+            (unsigned int)info.bpp, (unsigned int)info.pitch);
+    kprintf("Console grid: %ux%u cells\n", VGA_WIDTH, VGA_HEIGHT);
 }
