+++ src/shell/cmds/cmd_sha256sum.c
@@ -103,6 +103,6 @@ void cmd_sha256sum(const char *args) {
 
     uint8_t digest[32];
     sha256(buf, size, digest);
-    for (int i = 0; i < 32; i++) kprintf("%x%x", (unsigned long)(digest[i]>>4), (unsigned long)(digest[i]&0xF));
+    for (int i = 0; i < 32; i++) kprintf("%x%x", (unsigned int)(digest[i]>>4), (unsigned int)(digest[i]&0xF));
     kprintf("  %s\n", path);
 }
