+++ src/shell/cmds/cmd_unxz.c
@@ -74,5 +74,5 @@ done_blocks:
     }
 
     libc_vfs_write(outpath, outbuf, opos);
-    kprintf("Decompressed %u -> %u bytes (%s)\n", (unsigned long)in_size, (unsigned long)opos, outpath);
+    kprintf("Decompressed %llu -> %llu bytes (%s)\n", (unsigned long long)in_size, (unsigned long long)opos, outpath);
 }
